// projects/zweihander-1e/enums.ts
export enum AttributeId {
  COMBAT = 'attribute:combat',
  BRAWN = 'attribute:brawn',
  AGILITY = 'attribute:agility',
  PERCEPTION = 'attribute:perception',
  INTELLIGENCE = 'attribute:intelligence',
  WILLPOWER = 'attribute:willpower',
  FELLOWSHIP = 'attribute:fellowship'
}

export enum DrawbackId {
  BAD_TICKER = 'drawback:bad_ticker',
  BLACK_CATARACT = 'drawback:black_cataract',
  BLEEDER = 'drawback:bleeder',
  BRANDED = 'drawback:branded',
  CHOLERIC_TEMPERAMENT = 'drawback:choleric_temperament',
  CROP_EAR = 'drawback:crop_ear',
  CURSED = 'drawback:cursed',
  DEAL_WITH_THE_DEVIL = 'drawback:deal_with_the_devil',
  DEBTRIDDEN = 'drawback:debtridden',
  DUNDERHEAD = 'drawback:dunderhead',
  EUNUCH = 'drawback:eunuch',
  LILYLIVERED = 'drawback:lilylivered',
  MELANCHOLIC_TEMPERAMENT = 'drawback:melancholic_temperament',
  NEER_DO_WELL = 'drawback:neer_do_well',
  NEMESIS = 'drawback:nemesis',
  PAINKILLER = 'drawback:painkiller',
  PERSECUTION_COMPLEX = 'drawback:persecution_complex',
  PHLEGMATIC_TEMPERAMENT = 'drawback:phlegmatic_temperament',
  SANGUINE_TEMPERAMENT = 'drawback:sanguine_temperament',
  SOUR_STOMACH = 'drawback:sour_stomach',
  SPLIT_FACE = 'drawback:split_face',
  VETERANS_BOOT = 'drawback:veterans_boot',
  VETERANS_EYE = 'drawback:veterans_eye',
  VETERANS_HAND = 'drawback:veterans_hand',
  VETERANS_LEG = 'drawback:veterans_leg',
  WEAK_LUNGS = 'drawback:weak_lungs',
  //
  LURE_OF_POWER = 'drawback:lure_of_power',
  CRUEL_TUTELAGE = 'drawback:cruel_tutelage',
  THE_END_IS_NIGH = 'drawback:the_end_is_nigh',
  WHEN_THE_DEALINS_DONE = 'drawback:when_the_dealins_done',
  CRESTFALLEN = 'drawback:crestfallen',
  ABSTAIN_FROM_VIOLENCE = 'drawback:abstain_from_violence',
  SALT_OF_THE_EARTH = 'drawback:salt_of_the_earth',
  NOTORIOUS_MOOCHER = 'drawback:notorious_moocher',
  WASHED_UP = 'drawback:washed_up',
  DEATH_WISH = 'drawback:death_wish',
  ALTER_EGO = 'drawback:alter_ego',
  WARP_DISSONANCE = 'drawback:warp_dissonance',
  ASSASSINS_CREED = 'drawback:assassins_creed',
  THE_CUSTODIANS_COMMANDMENTS = 'drawback:the_custodians_commandments',
  ENJOY_THE_SILENCE = 'drawback:enjoy_the_silence',
  // LURE_OF_POWER = 'drawback:aaaa',
  // LURE_OF_POWER = 'drawback:aaaa',
  // LURE_OF_POWER = 'drawback:aaaa',
  // LURE_OF_POWER = 'drawback:aaaa',
}

export enum ProfessionId {
  ADHERENT = 'profession:adherent',
  ANARCHIST = 'profession:anarchist',
  ANCHORITE = 'profession:anchorite',
  ANIMAL_TAMER = 'profession:animal_tamer',
  ANTIQUARIAN = 'profession:antiquarian',
  APOTHECARY = 'profession:apothecary',
  ARTISAN = 'profession:artisan',
  ASTROLOGER = 'profession:astrologer',
  BAILIFF = 'profession:bailiff',
  BARBER_SURGEON = 'profession:barber_surgeon',
  // ADHERENT = 'profession:adherent',
  // ADHERENT = 'profession:adherent',
  // ADHERENT = 'profession:adherent',
  // ADHERENT = 'profession:adherent',
  // ADHERENT = 'profession:adherent',
  // ADHERENT = 'profession:adherent',
  // ADHERENT = 'profession:adherent',
  // ADHERENT = 'profession:adherent',
  // ADHERENT = 'profession:adherent',
  // ADHERENT = 'profession:adherent',
  // ADHERENT = 'profession:adherent',
  // ADHERENT = 'profession:adherent',
  // ADHERENT = 'profession:adherent',
  // ADHERENT = 'profession:adherent',
  // ADHERENT = 'profession:adherent',
  // ADHERENT = 'profession:adherent',
  // ADHERENT = 'profession:adherent',
  // ADHERENT = 'profession:adherent',
  // ADHERENT = 'profession:adherent',
  // ADHERENT = 'profession:adherent',
  // ADHERENT = 'profession:adherent',
  // ADHERENT = 'profession:adherent',
  // ADHERENT = 'profession:adherent',
  // ADHERENT = 'profession:adherent',
  // ADHERENT = 'profession:adherent',
  // ADHERENT = 'profession:adherent',
  // ADHERENT = 'profession:adherent',
  // ADHERENT = 'profession:adherent',
  // ADHERENT = 'profession:adherent',
  // ADHERENT = 'profession:adherent',
  // ADHERENT = 'profession:adherent',
}

export enum SkillId {
  ALCHEMY = 'skill:alchemy',
  ATHLETICS = 'skill:athletics',
  AWARENESS = 'skill:awareness',
  BARGAIN = 'skill:bargain',
  CHARM = 'skill:charm',
  COORDINATION = 'skill:coordination',
  COUNTERFEIT = 'skill:counterfeit',
  DISGUISE = 'skill:disguise',
  DRIVE = 'skill:drive',
  EAVESDROP = 'skill:eavesdrop',
  EDUCATION = 'skill:education',
  FOLKLORE = 'skill:folklore',
  GAMBLE = 'skill:gamble',
  GUILE = 'skill:guile',
  HANDLE_ANIMAL = 'skill:handle_animal',
  HEAL = 'skill:heal',
  INCANTATION = 'skill:incantation',
  INTERROGATION = 'skill:interrogation',
  INTIMIDATE = 'skill:intimidate',
  LEADERSHIP = 'skill:leadership',
  MARTIAL_MELEE = 'skill:martial_melee',
  MARTIAL_RANGED = 'skill:martial_ranged',
  NAVIGATION = 'skill:navigation',
  PILOT = 'skill:pilot',
  RESOLVE = 'skill:resolve',
  RIDE = 'skill:ride',
  RUMOR = 'skill:rumor',
  SCRUTINIZE = 'skill:scrutinize',
  SIMPLE_MELEE = 'skill:simple_melee',
  SIMPLE_RANGED = 'skill:simple_ranged',
  SKULDUGGERY = 'skill:skulduggery',
  STEALTH = 'skill:stealth',
  SURVIVAL = 'skill:survival',
  TOUGHNESS = 'skill:toughness',
  TRADECRAFT = 'skill:tradecraft',
  WARFARE = 'skill:warfare'
}

export enum TalentId {
  AETHEREAL_ALIGNMENT = 'talent:aethereal_alignment',
  AMBIDEXTERITY = 'talent:ambidexterity',
  APPALLING_MIEN = 'talent:appalling_mien',
  ARBALESTS_SPEED = 'talent:arbalests_speed',
  AZIMUTH = 'talent:azimuth',
  BAD_AXX = 'talent:bad_axx',
  BATTLE_MAGICK = 'talent:battle_magick',
  BEATDOWN = 'talent:beatdown',
  BLOOD_MAGICK = 'talent:blood_magick',
  CAROUSING = 'talent:carousing',
  CHEAP_SHOT = 'talent:cheap_shot',
  CLINCH_FIGHTER = 'talent:clinch_fighter',
  DETERMINATION = 'talent:determination',
  DIE_HARD = 'talent:die_hard',
  DOPPELGANGER = 'talent:doppelganger',
  EAGLE_EYES = 'talent:eagle_eyes',
  FENCERS_PANACHE = 'talent:fencers_panache',
  FORKED_TONGUE = 'talent:forked_tongue',
  GALLOWS_HUMOR = 'talent:gallows_humor',
  GANGSTER_GRIP = 'talent:gangster_grip',
  GATECRASHER = 'talent:gatecrasher',
  GROUND_POUND = 'talent:ground_pound',
  GRUESOME_SHOT = 'talent:gruesome_shot',
  HANDSPRING = 'talent:handspring',
  HARD_TO_KILL = 'talent:hard_to_kill',
  HIGHER_MYSTERIES = 'talent:higher_mysteries',
  HOLDOUT = 'talent:holdout',
  HOUSEBREAKER = 'talent:housebreaker',
  IMPENETRABLE_WALL = 'talent:impenetrable_wall',
  IMPERVIOUS_MIND = 'talent:impervious_mind',
  INDIFFERENCE = 'talent:indifference',
  INCREDIBLE_NUMERATION = 'talent:incredible_numeration',
  INSTINCTS = 'talent:instincts',
  KIDNEY_SHOT = 'talent:kidney_shot',
  KNIFEWORK = 'talent:knifework',
  LARCENY = 'talent:larceny',
  LEFT_HANDED_PATH = 'talent:left_handed_path',
  LIGHTNING_REACTION = 'talent:lightning_reaction',
  LIGHT_SLEEPER = 'talent:light_sleeper',
  LONG_WINDED = 'talent:long_winded',
  MARINER = 'talent:mariner',
  MENACING_DEMEANOR = 'talent:menacing_demeanor',
  MEETING_OF_THE_MINDS = 'talent:meeting_of_the_minds',
  MILITARY_FORMATION = 'talent:military_formation',
  MINE_CRAFT = 'talent:mine_craft',
  MOUNTED_DEFENSE = 'talent:mounted_defense',
  MULTILINGUAL = 'talent:multilingual',
  NERVES_OF_STEEL = 'talent:nerves_of_steel',
  NO_MERCY = 'talent:no_mercy',
  OVERWHELMING_FORCE = 'talent:overwhelming_force',
  RUN_AMOK = 'talent:run_amok',
  RURAL_SENSIBILITY = 'talent:rural_sensibility',
  SADDLE_BORN = 'talent:saddle_born',
  SECOND_SKIN = 'talent:second_skin',
  SECRET_SIGNS = 'talent:secret_signs',
  SHIELD_SLAM = 'talent:shield_slam',
  SHOOT_FROM_THE_HIP = 'talent:shoot_from_the_hip',
  SIEGECRAFT = 'talent:siegecraft',
  SILVER_TONGUE = 'talent:silver_tongue',
  SPIRITED_CHARGE = 'talent:spirited_charge',
  SPRINT = 'talent:sprint',
  STRANGLERS_UNION = 'talent:stranglers_union',
  STREETWISE = 'talent:streetwise',
  STRONG_JAW = 'talent:strong_jaw',
  SUPERNATURAL_PARANOIA = 'talent:supernatural_paranoia',
  SWORD_BOARD = 'talent:sword_board',
  TAKE_EM_DOWN = 'talent:take_em_down',
  THERE_WILL_BE_BLOOD = 'talent:there_will_be_blood',
  TOUGH_AS_NAILS = 'talent:tough_as_nails',
  TRUE_GRIT = 'talent:true_grit',
  WINDS_OF_CHANGE = 'talent:winds_of_change',
  WORLDLY = 'talent:worldly'
}